Changing to Detector Collection
